Rename resources to new project
