+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0]
+## [Unreleased]
+
+### Changes
+
+- Change package group from com.bisnode.kafka.authorization to org.openpolicyagent.kafka
+
+## [1.2.0] - 2021-10-12
+
+### Changes
+
+- Ensure compatibility with Kafka 3.0.0 (@scholzj)
+
+## [1.1.0] - 2021-06-11
 
 ### Changes
