+++ src/test/scala/org/openpolicyagent/kafka/OpaAuthorizerSpec.scala
@@ -1,37 +1,27 @@
-package com.bisnode.kafka.authorization
+package org.openpolicyagent.kafka
 
 import java.net.{InetAddress, URI}
 import java.net.http.HttpRequest.BodyPublishers
 import java.net.http.HttpResponse.BodyHandlers
 import java.net.http.{HttpClient, HttpRequest, HttpResponse}
-import java.util.Collections
-import java.{util => ju}
 import org.junit.runner.RunWith
 import org.scalatestplus.junit.JUnitRunner
 import com.typesafe.scalalogging.LazyLogging
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.scala.{DefaultScalaModule, ScalaObjectMapper}
 import kafka.network.RequestChannel
-import kafka.network.RequestChannel.Session
 import org.apache.kafka.common.acl.AclOperation
 import org.apache.kafka.common.network.{ClientInformation, ListenerName}
 import org.apache.kafka.common.protocol.ApiKeys
 import org.apache.kafka.common.resource.PatternType
 import org.apache.kafka.common.security.auth.{KafkaPrincipal, SecurityProtocol}
 import org.apache.kafka.common.resource.ResourcePattern
-import org.apache.kafka.common.resource.PatternType
 import org.apache.kafka.common.resource.ResourceType
 import org.apache.kafka.common.requests.{RequestContext, RequestHeader}
 import org.apache.kafka.server.authorizer.{Action, AuthorizationResult}
 import org.scalatest._
 import matchers.should._
 import flatspec._
-import org.apache.kafka.common.ClusterResource
-import org.apache.kafka.common.Endpoint
-import scala.compat.java8.FutureConverters
-import io.netty.util.concurrent.Future
-import scala.util.{Success, Failure}
-import scala.concurrent.ExecutionContext.Implicits.global
 import scala.jdk.CollectionConverters._
 
 /**
